Fixed /hl list, /hl clear, added limitations to highlight system
- /hl list now properly retrieves the entries from the database and displays them.
- /hl clear no longer allow you to clear multiple times in a row.
- Implemented 50 character limit to avoid embed max length overflows.
- Implemented 20 highlight term limit to avoid strain and to avoid embed max length overflows.
- (Mostly) scraped the cache system. There are still remnants that need to be converted over for this PR to actually work.
